added barline class and modified alter type
Added Barline class as an element of a measure
modified the Xmlparser.cs to include methods for getting barline values.
modified alter type from int to double to be able to receive values of Arabic accidentals such (quarter note)

+++ MusicXml/Domain/Pitch.cs
@@ -11,7 +11,7 @@ namespace MusicXml.Domain
 			Step = new Char();
 		}
 
-		public int Alter { get; internal set; }
+		public double Alter { get; internal set; }
 
 		public int Octave { get; internal set; }
